moved the timer to new file
